+++ Dekoter/Classes/JSONSerialization+Dekoter.swift
@@ -27,21 +27,34 @@ import Foundation
 
 extension JSONSerialization {
     
-    class func de_jsonObject<T: Koting>(with data: Data, options opt: JSONSerialization.ReadingOptions = []) throws -> T {
-        let object = try jsonObject(with: data, options: opt)
-        guard let dict = object as? [AnyHashable: Any] else {
-            throw NSError(domain: "Dekoter", code: 5001, userInfo: ["NSDebugDescription": "JSON object is not a dictionary."])
+    /// Creates an object which implements the `Koting` protocol from JSON data.
+    ///
+    /// - Parameters:
+    ///   - data: JSON data.
+    ///   - opt: Options.
+    /// - Returns: A deserialized object.
+    class func de_jsonObject<T: Koting>(with data: Data, options opt: JSONSerialization.ReadingOptions = []) -> T? {
+        guard let object = try? jsonObject(with: data, options: opt),
+            let dict = object as? [AnyHashable: Any] else {
+            
+            return nil
         }
-        return try de_serializedObject(from: dict)
+        return de_serializedObject(from: dict)
     }
     
-    class func de_jsonObject<T: Koting>(with data: Data, options opt: JSONSerialization.ReadingOptions = []) throws -> [T] {
-        let object = try jsonObject(with: data, options: opt)
-        guard let dicts = object as? [[AnyHashable: Any]] else {
-            throw NSError(domain: "Dekoter", code: 5003, userInfo: ["NSDebugDescription": "JSON object is not an array of dictionaries."])
+    /// Creates an array of objects which implement the `Koting` protocol from JSON data.
+    ///
+    /// - Parameters:
+    ///   - data: JSON data.
+    ///   - opt: Options.
+    /// - Returns: A deserialized array.
+    class func de_jsonObject<T: Koting>(with data: Data, options opt: JSONSerialization.ReadingOptions = []) -> [T]? {
+        guard let object = try? jsonObject(with: data, options: opt),
+            let dicts = object as? [[AnyHashable: Any]] else {
+            
+            return nil
         }
-        
-        return try dicts.map { try de_serializedObject(from: $0) }
+        return dicts.flatMap { de_serializedObject(from: $0) }
     }
 }
 
@@ -49,10 +62,10 @@ extension JSONSerialization {
 
 fileprivate extension JSONSerialization {
     
-    class func de_serializedObject<T: Koting>(from dict: [AnyHashable: Any]) throws -> T {
+    class func de_serializedObject<T: Koting>(from dict: [AnyHashable: Any]) -> T? {
         let coder = JSONKoter(objects: dict)
         guard let object = T(koter: coder) else {
-            throw NSError(domain: "Dekoter", code: 5002, userInfo: ["NSDebugDescription": "JSON object doesn't conform to the requested type."])
+            return nil
         }
         return object
     }